+++ connector/tests/test_job.py
@@ -15,6 +15,7 @@ from openerp.addons.connector.queue.job import (
     OpenERPJobStorage,
     job,
     _unpickle,
+    RETRY_INTERVAL,
 )
 from openerp.addons.connector.session import (
     ConnectorSession,
@@ -117,12 +118,80 @@ class TestJobs(unittest2.TestCase):
     def test_retryable_error(self):
         test_job = Job(func=retryable_error_task,
                        max_retries=3)
+        self.assertEqual(test_job.retry, 0)
         with self.assertRaises(RetryableJobError):
             test_job.perform(self.session)
+        self.assertEqual(test_job.retry, 1)
         with self.assertRaises(RetryableJobError):
             test_job.perform(self.session)
+        self.assertEqual(test_job.retry, 2)
         with self.assertRaises(FailedJobError):
             test_job.perform(self.session)
+        self.assertEqual(test_job.retry, 3)
+
+    def test_infinite_retryable_error(self):
+        test_job = Job(func=retryable_error_task,
+                       max_retries=0)
+        self.assertEqual(test_job.retry, 0)
+        with self.assertRaises(RetryableJobError):
+            test_job.perform(self.session)
+        self.assertEqual(test_job.retry, 1)
+
+    def test_retry_pattern(self):
+        """ When we specify a retry pattern, the eta must follow it"""
+        datetime_path = 'openerp.addons.connector.queue.job.datetime'
+        test_pattern = {
+            1: 60,
+            2: 180,
+            3: 10,
+            5: 300,
+        }
+        job(retryable_error_task, retry_pattern=test_pattern)
+        with mock.patch(datetime_path, autospec=True) as mock_datetime:
+            mock_datetime.now.return_value = datetime(2015, 6, 1, 15, 10, 0)
+            test_job = Job(func=retryable_error_task,
+                           max_retries=0)
+            test_job.retry += 1
+            test_job.postpone(self.session)
+            self.assertEqual(test_job.retry, 1)
+            self.assertEqual(test_job.eta, datetime(2015, 6, 1, 15, 11, 0))
+            test_job.retry += 1
+            test_job.postpone(self.session)
+            self.assertEqual(test_job.retry, 2)
+            self.assertEqual(test_job.eta, datetime(2015, 6, 1, 15, 13, 0))
+            test_job.retry += 1
+            test_job.postpone(self.session)
+            self.assertEqual(test_job.retry, 3)
+            self.assertEqual(test_job.eta, datetime(2015, 6, 1, 15, 10, 10))
+            test_job.retry += 1
+            test_job.postpone(self.session)
+            self.assertEqual(test_job.retry, 4)
+            self.assertEqual(test_job.eta, datetime(2015, 6, 1, 15, 10, 10))
+            test_job.retry += 1
+            test_job.postpone(self.session)
+            self.assertEqual(test_job.retry, 5)
+            self.assertEqual(test_job.eta, datetime(2015, 6, 1, 15, 15, 00))
+
+    def test_retry_pattern_no_zero(self):
+        """ When we specify a retry pattern without 0, uses RETRY_INTERVAL"""
+        test_pattern = {
+            3: 180,
+        }
+        job(retryable_error_task, retry_pattern=test_pattern)
+        test_job = Job(func=retryable_error_task,
+                       max_retries=0)
+        test_job.retry += 1
+        self.assertEqual(test_job.retry, 1)
+        self.assertEqual(test_job._get_retry_seconds(), RETRY_INTERVAL)
+        test_job.retry += 1
+        self.assertEqual(test_job.retry, 2)
+        self.assertEqual(test_job._get_retry_seconds(), RETRY_INTERVAL)
+        test_job.retry += 1
+        self.assertEqual(test_job.retry, 3)
+        self.assertEqual(test_job._get_retry_seconds(), 180)
+        test_job.retry += 1
+        self.assertEqual(test_job.retry, 4)
+        self.assertEqual(test_job._get_retry_seconds(), 180)
 
     def test_unpickle(self):
         pickle = ("S'a small cucumber preserved in vinegar, "
@@ -508,3 +577,13 @@ class TestJobChannels(common.TransactionCase):
         self.assertEquals(channel.name, 'subsub')
         self.assertEquals(channel.parent_id.name, 'sub')
         self.assertEquals(channel.parent_id.parent_id.name, 'root')
+
+    def test_job_decorator(self):
+        """ Test the job decorator """
+        default_channel = 'channel'
+        retry_pattern = {1: 5}
+        partial = job(None, default_channel=default_channel,
+                      retry_pattern=retry_pattern)
+        self.assertEquals(partial.keywords.get('default_channel'),
+                          default_channel)
+        self.assertEquals(partial.keywords.get('retry_pattern'), retry_pattern)